Factor out code responsible for gathering files in a directory.
