+++ pdns/dnsdistdist/logger.hh
@@ -1,0 +1,1 @@
+../logger.hh
