+++ pdns/misc.hh
@@ -45,6 +45,7 @@
 #include <vector>
 
 #include "namespaces.hh"
+#include "logr.hh"
 
 class DNSName;
 #if defined(PDNS_AUTH)
@@ -975,6 +976,8 @@ private:
 namespace pdns
 {
 [[nodiscard]] std::optional<std::string> visit_directory(const std::string& directory, const std::function<bool(ino_t inodeNumber, const std::string_view& name)>& visitor);
+
+std::vector<std::string> list_directory(const std::string& directory, const std::string& suffix, Logr::log_t d_log);
 
 struct FilePtrDeleter
 {
